[IMP] im_livechat: introduce LivechatButtonView/willStart (part 1)
Task-2937815
